% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Run On Load}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{(as required)}

\item{pkgname}{(as required)}
}
\description{
This initialises the user environment when the package is attached to make sure
that required git username / password is available (and notify which is being
used). It is highly anticipated that this will be used in interactive mode,
and may be reloaded several times in one session using \code{library(git4r)}
}
\details{
Some of the environmental options can be set as .Renviron variables, such
as username, local remote directory, and personal access token secrets.
}
