% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Run On Load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{(as required)}

\item{pkgname}{(as required)}
}
\description{
This initialises the user environment when the package is loaded to make sure
that required git username / password is available (and notify which is being
used).
}
\details{
Some of the environmental options can be set as .Renviron variables, such
as username, local remote directory, and personal access token secrets.
}
