% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_utils.R
\name{edit_tempfile}
\alias{edit_tempfile}
\title{Edit File In Tempdir()}
\usage{
edit_tempfile(filename)
}
\arguments{
\item{filename}{Character vector of filenames which exist in \code{tempdir()} to
open for editing. Parses with \code{basename()} so no subdirectories allowed.}
}
\description{
\code{git_diff()} saves copies of each file opened in \code{tempdir()} which can be
opened for editing by this helper.
}
