% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_utils.R
\name{fix_path}
\alias{fix_path}
\title{Clean File Path}
\usage{
fix_path()
}
\value{
Atomic character vector giving the cleaned path string
}
\description{
Converts backslashes to forward and removes and double-slash except a possible
double-slash at the start for a UNC path. Paste in the dirty path interactively
to allow single-backslash-paths to be pasted in without issues
}
