% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit.R
\name{git}
\alias{git}
\title{Auto Git}
\usage{
git()
}
\value{
Invisible NULL
}
\description{
Call the git pull-add-commit-push cycle interactively. This is the default
helper function to call continuously to log changes and stay up-to-date.
}
\seealso{
git_remote, git_add, git_commit, git_push, git_branch, git_diff
}
