% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_branch.R
\name{git_branch}
\alias{git_branch}
\title{Git Branch}
\usage{
git_branch(branchname = NULL, deletebranch = NULL)
}
\arguments{
\item{branchname}{Branch name to move to or create, use empty string "" to list
existing branches}

\item{deletebranch}{If moving off a branch which has no commits on it, should the departed
branch be deleted? "Y" or TRUE, else FALSE}
}
\description{
Swaps to a different branch if exists, otherwise makes one from most recent
commit, and checks out this branch. Hit ENTER on first question to see a list
of available branches.
}
\details{
If there are uncommitted changes, it will forbid changing to an existing
branch which would result in irreversible loss of current working directory.
If the branch does not exist yet, the working directory is not changed and
the next commit will be on this new branch.

You are allowed to change branch to any which share the same latest commit
so you can create a branch and instantly remove it, whilst keeping any
pending changes.

Changing to a branch which is only listed under remote (for example origin/test)
by \code{git_branch('test')} will automatically create a local copy of this branch.
}
\seealso{
git_merge
}
