% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit.R
\name{git_commit}
\alias{git_commit}
\title{Git Commit}
\usage{
git_commit(message = NULL)
}
\arguments{
\item{message}{Commit message, usually one sentence about a specific change,
character atomic. This is asked for interactively if left NULL.}
}
\value{
Invisible NULL
}
\description{
Run git_add() immediately before to see what files are being committed. It is
highly recommended that you run in interactive mode using no arguments
(\code{git_commit()})
}
