% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_remote.R
\name{git_pull}
\alias{git_pull}
\title{Git Pull}
\usage{
git_pull()
}
\value{
Invisible NULL
}
\description{
This will pull any updates from the 'origin' remote for every branch.
This should be run before making a commit (as is done by wrapper \code{git()}) in
order to avoid conflicts arising if somebody else has made changes
on the same branch. If this happens, merge conflicts are raised in the usual
way and \code{git_add()} will show you what needs resolving.
}
\details{
To setup an 'origin' remote, use \code{git_remote()}. Credentials are handled
in the same way as \code{git_push()}, see \code{?git_push} for details.
}
\seealso{
git_push, git_remote
}
