% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_utils.R
\name{xcopy}
\alias{xcopy}
\alias{xpaste}
\title{Copy and Paste to Excel}
\usage{
xcopy(x, header = TRUE)

xpaste(header = TRUE)
}
\arguments{
\item{x}{A data.frame or any table which can be written with \code{write.table()}}

\item{header}{Data includes headers -- the only reason to omit when pasting is
concatenating multiple chunks below each other}
}
\description{
This reads a table from a (enlarged) clipboard or writes as tab-separated
which behaves nicely in Excel
}
\details{
Pasting from Excel breaks if the text contains linebreaks within a cell.
To fix this, use Find&Replace "ctrl+J" within Excel

If \code{data.table} is attached, xpaste() will return a data.table
}
