% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_remote.R
\name{git_clone}
\alias{git_clone}
\title{Git Clone from Default Remote}
\usage{
git_clone(repo_name = "", to = "~")
}
\arguments{
\item{repo_name}{Name of repo to clone, leave blank to display a list of available
repo names. Also accepted is the url or complete filepath to a
repo to clone (identified if containing '/').}

\item{to}{Local path to create local project folder in, defaults to My Documents}
}
\description{
Running without any repo_name will print a list of valid repos found at the
GIT_DEFAULT_REMOTE path, and can then be re-run with one of these names to
clone a local version. Otherwise a url or complete filepath can be used.
}
\details{
To change where the default remote path is use
\itemize{
\item \code{Sys.setenv(GIT_DEFAULT_REMOTE='P:/ath/to/teams/remote_repo/folder')}
}
}
