% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit.R
\name{git_commit}
\alias{git_commit}
\title{Git Commit}
\usage{
git_commit(message = NULL, proceed = NULL)
}
\arguments{
\item{message}{Commit message, usually one sentence about a specific change,
character atomic.}

\item{proceed}{Confirmation before proceeding, logical atomic with option for
any character string starting 'Y' to be interpreted as \code{TRUE}}
}
\description{
Run git_add() immediately before to see what files are being committed. It is
highly recommended that you run in interactive mode using no arguments
(\code{git_commit()})
}
