% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_branch.R
\name{git_merge}
\alias{git_merge}
\title{Merge Branch into Current Branch}
\usage{
git_merge(branchname = NULL)
}
\arguments{
\item{branchname}{Name of other branch to merge into the current one, will be asked
interactively if left as NULL.}
}
\value{
Invisible NULL
}
\description{
For good house-keeping, the merged branch is by default deleted, this can be
recreated any time using git_branch(). Conflict resolution is yet to be
implemented, for now the most recent of a file is kept unless changes have
happened in both branches.
}
\details{
A really neat tool in RStudio is \code{Edit} -> \verb{Find in Files...} which allows
you to search your entire repository for where the conflicts are. These are
identified by searching for the chevrons <<<

Conflicts are the usual format for git and may also happen after git_pull().

\preformatted{
<<<<<<< HEAD
code-from-the-branch-you-have-stayed-on
=======
code-from-the-branch-you-have-just-merged
>>>>>>> merged-branch-name
}
}
