% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_remote.R
\name{git_pull}
\alias{git_pull}
\title{Git Pull}
\usage{
git_pull()
}
\value{
Invisible NULL
}
\description{
This will pull any updates from the remote which each branch is tracking.
This is set using git_remote() to be the 'origin' remote, usually at
GIT_DEFAULT_REMOTE. This should be run before any new commits.
}
\details{
If git_push() has not been run yet, it will complain that branches are not
set to track anything. If so, run git_remote() and git_push() to synchronise.

Credentials are sorted in the same way as git_pull() and git_push()
using the best-matching \code{GIT_PAT}* environmental variable to the remote's
address, for example \verb{GIT_PAT_AZURE=} in \verb{~/.Renviron}
}
\seealso{
git_push, git_remote
}
