% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_remote.R
\name{git_push}
\alias{git_push}
\title{Git Push}
\usage{
git_push(do_default = NULL, remote_index = NULL, branch_index = NULL)
}
\arguments{
\item{do_default}{Confirm that the default push-all-tracking-branches-to-origin ("Y" or TRUE)}

\item{remote_index}{If this is not true, selection of remote to push to (integer vector)}

\item{branch_index}{The selection of branches to push (integer vector)}
}
\description{
Pushes just current branch to 'origin' by default, but allows selective
pushing of branches to different remotes if this is declined, for example
if you want to keep a test branch private.
}
\details{
It is enforced that when pushing to 'origin' the branch is set to track this
remote branch. If the branch does not exist yet in the remote, it is not
pushed by default and must be manually added.

Credentials are sorted in the same way as git_pull() and git_push()
using the best-matching \verb{GIT_PAT*} environmental variable to the remote's
address, for example \verb{GIT_PAT_AZURE=} in \verb{~/.Renviron}
}
\seealso{
git_pull, git_remote
}
